Convert login flow to function in Utils class
